backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
aiohttp

fastmcp>=0.5.0
mcp[cli]
python-dotenv

pytest
pytest-asyncio
